Break the basic memory model data structures into a separate module
